requirements.txt
alembic==0.8.10
beautifulsoup4==4.5.3
blinker==1.4
Flask==0.12
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
Jinja2==2.8.1
MySQL-python==1.2.5
Pillow==4.0.0
python-editor==1.0.3
python-memcached==1.58
qiniu==7.1.2
redis==2.10.5
requests==2.13.0
requests-toolbelt==0.7.1
sendEMail==1.0
shortuuid==0.5.0
six==1.10.0
SQLAlchemy==1.1.5
Top==1.0.0
virtualenv==15.1.0
visitor==0.1.3
Werkzeug==0.11.15
WTForms==2.1
